+++ gpt_disk_types/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Documentation improvements.
+
 # 0.15.0
 
 * Updated to latest `uguid`.